refactor app class
